+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - Add `Fiber` integration ([#795](https://github.com/getsentry/sentry-go/pull/795))
 - Use `errors.Unwrap()` to create exception groups ([#792](https://github.com/getsentry/sentry-go/pull/792))
 
+### Fixes
+
+- Fix missing stack trace for parsing error in logrusentry ([#689](https://github.com/getsentry/sentry-go/pull/689))
+
 ## 0.27.0
 
 The Sentry SDK team is happy to announce the immediate availability of Sentry Go SDK v0.27.0.